refactor: extract big block to another file
